Harden against accidental leaking of the password
In order to prevent submission of the password form of the
self-decrypting HTML page from causing the password from being sent over
the network, the JavaScript method preventDefault() was used.
However, this fails to take effect if the browser does not have
JavaScript enabled, and could ultimately lead to the password being
leaked to the network, which indeed happens with Lynx 2.8.9rel.1.

Hence set the action attribute explicitly to prevent the default action
from sending anything over the network. Also hide the form and display
a helpful message instead if the browser does not have JavaScript
enabled.